#312 additional fixes
